Created the payment jwt and orders db

+++ src/routes/admin/admin.py
@@ -1,13 +1,20 @@
-from flask import Blueprint, request
+from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt
 import json
 
 
 from src.db.products_db import insertProduct, updateProductById, deleteProduct
+from src.db.orders_db import getAllOrders
 
 admin = Blueprint("admin", __name__)
 
 @admin.route("/addProduct", methods = ["POST"])
+@jwt_required()
 def addProduct():
+
+    claims = get_jwt()
+    if claims['role'] != 'admin':
+        return jsonify({"msg": "Access forbidden: Admins only!"}), 403
 
     if 'image' not in request.files:
         return json.dumps({"error":"No image available"}), 400
@@ -30,7 +37,12 @@ def addProduct():
         print(e)
 
 @admin.route("/updateProduct/<id>", methods = ["POST"])
+@jwt_required()
 def updateProduct(id):
+
+    claims = get_jwt()
+    if claims['role'] != 'admin':
+        return jsonify({"msg": "Access forbidden: Admins only!"}), 403
 
     if 'image' not in request.files:
         return json.dumps({"error":"No image available"}), 400
@@ -53,8 +65,24 @@ def updateProduct(id):
         print(e)
 
 @admin.route("/deleteProduct/<id>")
+@jwt_required()
 def deleteProductById(id):
-      return deleteProduct(id)
+    claims = get_jwt()
+    if claims['role'] != 'admin':
+        return jsonify({"msg": "Access forbidden: Admins only!"}), 403
+    
+    return deleteProduct(id)
+
+
+@admin.route("/getAllOrders", methods = ["POST"])
+@jwt_required()
+def getOrders():
+
+    claims = get_jwt()
+    if claims['role'] != 'admin':
+        return jsonify({"msg": "Access forbidden: Admin only!"}), 403
+
+    return getAllOrders()
 
 @admin.route("/viewCustomers")
 def viewCustomers():
